Port / move tests/models/test_user.py
